Printable Version Base Plate
The design calls for this piece to be cut in acrylic. I've cut the piece into two pieces then created overlap so that there is some strength where the pieces come together. Printed with ABS at 100% infill this seems to work. Won't know for sure until the robot is complete and have a chance to see how it holds up. May revised the design with additional support if needed. I prefer to print as much as possible not just for cost but also replacement parts.