Fan holders with built in hinge pins
The original design printed the fan holders with the assumption that another piece would be used to form the hinge. I am trying this to see if printing with the pins built in will fit nicely and hold up. My print is being done with ABS and 100% infill. Please use at your own risk. These may fail.