Top plate of robot split in half with overlap
My printer was not large enough to print this in one piece. I modified the part to split it in half then overlap tabs on each side for a strong connection between the two pieces. If you want to glue you can apply the glue to the tabs so that you get good adhesion in the middle of the two parts. Then after they dry more glue can be applied to the seems so that it is a very strong part.